requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

attrs==24.2.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via trio
charset-normalizer==3.3.2
    # via requests
h11==0.14.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via webdriver-manager
pillow==10.4.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pysocks==1.7.1
    # via urllib3
python-dotenv==1.0.1
    # via webdriver-manager
requests==2.32.3
    # via webdriver-manager
ruff==0.6.8
    # via -r requirements.in
selenium==4.25.0
    # via -r requirements.in
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.6
    # via beautifulsoup4
trio==0.26.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via selenium
urllib3[socks]==2.2.3
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r requirements.in
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
